+++ tests/test_reattachment.py
@@ -170,7 +170,7 @@ def plasma_volume(
     areal_elongation,
     inverse_aspect_ratio,
 ):
-    v = formulas.geometry.calc_plasma_volume(
+    v = formulas.geometry.analytical.calc_plasma_volume(
         major_radius=major_radius,
         areal_elongation=areal_elongation,
         inverse_aspect_ratio=inverse_aspect_ratio,
@@ -230,7 +230,7 @@ def two_point_model_fixed_tet(
     SOL_momentum_loss_function,
     sheath_heat_transmission_factor,
 ):
-    Te_tar = formulas.scrape_off_layer.two_point_model_fixed_tet(
+    Te_tar = formulas.scrape_off_layer.two_point_model.two_point_model_algorithms.two_point_model_fixed_tet(
         target_electron_temp=target_electron_temp,
         q_parallel=q_parallel,
         parallel_connection_length=parallel_connection_length,
@@ -250,7 +250,7 @@ def neutral_flux_density_factor(
     ratio_of_molecular_to_ion_mass,
     wall_temperature,
 ):
-    factor = formulas.scrape_off_layer.calc_neutral_flux_density_factor(
+    factor = formulas.scrape_off_layer.reattachment_models.calc_neutral_flux_density_factor(
         average_ion_mass=average_ion_mass,
         ratio_of_molecular_to_ion_mass=ratio_of_molecular_to_ion_mass,
         wall_temperature=wall_temperature,
@@ -276,7 +276,7 @@ def target_neutral_pressure(
     q_parallel,
     ratio_of_divertor_to_duct_pressure,
 ):
-    p_div, p_duct = formulas.scrape_off_layer.calc_neutral_pressure_kallenbach(
+    p_div, p_duct = formulas.scrape_off_layer.reattachment_models.calc_neutral_pressure_kallenbach(
         average_ion_mass=average_ion_mass,
         kappa_e0=kappa_e0,
         kappa_ez=kappa_ez,
@@ -301,7 +301,7 @@ def target_neutral_pressure(
 def ionization_volume(
     plasma_volume,
 ):
-    ionization_volume = formulas.scrape_off_layer.calc_ionization_volume_from_AUG(plasma_volume=plasma_volume)
+    ionization_volume = formulas.scrape_off_layer.reattachment_models.calc_ionization_volume_from_AUG(plasma_volume=plasma_volume)
     return ionization_volume
 
 
@@ -313,7 +313,7 @@ def test_calc_reattachment_time_henderson(
     ionization_volume_density_factor,
     ionization_volume,
 ):
-    reattachment_time = formulas.scrape_off_layer.calc_reattachment_time_henderson(
+    reattachment_time = formulas.scrape_off_layer.reattachment_models.calc_reattachment_time_henderson(
         target_neutral_pressure=target_neutral_pressure,
         target_electron_density=target_electron_density,
         parallel_connection_length=parallel_connection_length,